Refactor: removed all files and mentions from estacao pirata
