+++ Content.Server/Content.Server.csproj
@@ -25,6 +25,9 @@
     <ProjectReference Include="..\RobustToolbox\Robust.Server\Robust.Server.csproj" />
     <ProjectReference Include="..\Content.Shared\Content.Shared.csproj" />
   </ItemGroup>
+  <ItemGroup>
+    <Compile Remove="EstacaoPirata\Changeling\EntitySystems\ChangelingSystem.Abilities.cs" />
+  </ItemGroup>
 
   <Import Project="..\RobustToolbox\MSBuild\Robust.Properties.targets" />
 </Project>
