--- Content.Server/EstacaoPirata/Changeling/Components/AbsorbComponent.cs
@@ -1,11 +1,0 @@
-namespace Content.Server.EstacaoPirata.Changeling.Components;
-
-[RegisterComponent]
-public sealed class AbsorbComponent : Component
-{
-    /// <summary>
-    /// Whether or not the entity has been absorbed yet.
-    /// </summary>
-    [ViewVariables(VVAccess.ReadWrite)]
-    public bool Absorbed = false;
-}
