+++ Resources/Locale/en-US/actions/actions/changeling.ftl
@@ -1,0 +1,18 @@
+action-absorb-dna-name = Absorb DNA
+action-absorb-dna-description = Absorbs all the DNA of the target.
+
+action-arm-blade-name = Arm Blade
+action-arm-blade-description = Transforms one of your arms into a blade.
+
+action-dna-sting-name = DNA Sting
+action-dna-sting-description = Collects the DNA of the target.
+
+action-fleshmend-name = Fleshmend
+action-fleshmend-description = Heal yourself with time.
+
+action-transform-name = Transform
+action-transform-description = Change your physical structure to that of another person.
+
+changeling-dna-obtained = You obtain {THE($target)}'s DNA.
+changeling-dna-failed-nonHumanoid = This thing isn't humanoid.
+changeling-dna-failed-noDna = This thing does not have DNA.
