+++ Resources/Locale/en-US/changeling/changeling.ftl
@@ -1,0 +1,7 @@
+#items
+true-arm-blade-name = arm blade
+true-arm-blade-description = A grotesque blade made out of bone and flesh that cleaves through people as a hot knife through butter.
+
+#UI
+armblade-listing-name = Arm Blade
+armblade-listing-description = Transform one of our arms into a grotesque blade made of bone and flesh. It can be retracted using the ability again. The arm blade can also be used to forcefully open doors.
