+++ Resources/Locale/en-US/store/currency.ftl
@@ -8,3 +8,5 @@ store-currency-display-debugdollar = {$amount ->
 }
 store-currency-display-telecrystal = TC
 store-currency-display-stolen-essence = Stolen Essence
+
+store-currency-display-points = Points
