--- Resources/Locale/pt-BR/estacao-pirata/actions/actions/changeling.ftl
@@ -1,18 +1,0 @@
-action-absorb-dna-name = Absorver DNA
-action-absorb-dna-description = Absorve todo DNA do alvo.
-
-action-arm-blade-name = Lâmina de Braço
-action-arm-blade-description = Transforma o braço em uma lâmina.
-
-action-dna-sting-name = Picada de DNA
-action-dna-sting-description = Coleta o DNA do alvo.
-
-action-fleshmend-name = Cura de Carne
-action-fleshmend-description = Se cura com o tempo.
-
-action-transform-name = Transformar
-action-transform-description = Altera sua estrutura física para a de outra pessoa.
-
-changeling-dna-obtained = Você obtém o DNA d{THE($target)}.
-changeling-dna-failed-nonHumanoid = Esta criatura não é humanoide.
-changeling-dna-failed-noDna = Esta criatura não possui DNA.
