--- Resources/Locale/pt-BR/estacao-pirata/catalog/changeling.ftl
@@ -1,2 +1,0 @@
-armblade-listing-name = Lâmina de Braço
-armblade-listing-description = Transforma um dos nossos braços em uma lâmina grotesca feita de osso e carne. Ela pode ser recolhida usando a habilidade novamente. A lâmina de braço também pode ser usada para forçar a abertura de portas.
