--- Resources/Locale/pt-BR/estacao-pirata/entities/objects/weapons/melee/melee.ftl
@@ -1,2 +1,0 @@
-ent-TrueArmBlade = Lâmina de Braço
-    .desc = Uma lâmina grotesca feita de osso e carne que corta as pessoas como uma faca quente corta a manteiga.
