--- Resources/Locale/pt-BR/estacao-pirata/store/currency.ftl
@@ -1,1 +1,0 @@
-store-currency-display-points = Pontos
