--- Resources/Locale/pt-BR/estacao-pirata/store/store.ftl
@@ -1,2 +1,0 @@
-activate-changeling-shop-action-name = Menu de evolução
-activate-changeling-shop-action-description = Escolha as habilidades que você terá.
