+++ Resources/Prototypes/Actions/changeling.yml
@@ -1,0 +1,41 @@
+- type: instantAction
+  id: ChangelingShop
+  icon: Interface/Actions/shop.png
+  name: Shop
+  description: Opens the evolution shop.
+  serverEvent: !type:ChangelingShopActionEvent
+
+- type: entityTargetAction
+  id: AbsorbDNA
+  icon: { sprite: Interface/Actions/ling_actions.rsi, state: ling_absorb }
+  name: action-absorb-dna-name
+  description: action-absorb-dna-description
+  serverEvent: !type:AbsorbDNAActionEvent
+  
+- type: instantAction
+  id: ChangelingArmBlade
+  icon: { sprite: Interface/Actions/ling_actions.rsi, state: arm_blade }
+  name: action-arm-blade-name
+  description: action-arm-blade-description
+  serverEvent: !type:ChangelingArmBladeEvent
+
+- type: entityTargetAction
+  id: ChangelingDnaSting
+  icon: { sprite: Interface/Actions/ling_actions.rsi, state: dna_sting }
+  name: action-dna-sting-name
+  description: action-dna-sting-description
+  serverEvent: !type:ChangelingDnaStingEvent
+
+- type: instantAction
+  id: ChangelingFleshmend
+  icon: { sprite: Interface/Actions/ling_actions.rsi, state: fleshmend }
+  name: action-fleshmend-name
+  description: action-fleshmend-description
+  serverEvent: !type:ChangelingFleshmendEvent
+
+- type: instantAction
+  id: ChangelingTransform
+  icon: { sprite: Interface/Actions/ling_actions.rsi, state: transform }
+  name: action-transform-name
+  description: action-transform-description
+  serverEvent: !type:ChangelingTransformEvent
