+++ Resources/Prototypes/Actions/types.yml
@@ -148,3 +148,12 @@
   icon: { sprite: Clothing/Head/Hats/pyjamasyndicatered.rsi, state: icon }
   checkCanInteract: false
   event: !type:WakeActionEvent
+
+- type: instantAction
+  id: ActivateChangelingShop
+  name: activate-changeling-shop-action-name
+  description: activate-changeling-shop-action-description
+  checkCanInteract: false
+  itemIconStyle: BigAction
+  icon: Interface/Actions/shop.png
+  event: !type:ChangelingShopActionEvent
