+++ Resources/Prototypes/Entities/Objects/Weapons/Melee/truearmblade.yml
@@ -1,0 +1,23 @@
+- type: entity
+  name: arm-blade-name
+  parent: BaseItem
+  id: TrueArmBlade
+  description: arm-blade-description
+  components:
+  - type: Sharp
+  - type: Sprite
+    sprite: Objects/Weapons/Melee/armblade.rsi
+    state: icon
+  - type: MeleeWeapon
+    attackRate: 1
+    damage:
+      types:
+        Slash: 20
+        Piercing: 10
+  - type: Item
+    size: 15
+    sprite: Objects/Weapons/Melee/armblade.rsi
+  - type: Tool
+    qualities:
+      - Prying
+  - type: Unremoveable
