--- Resources/Prototypes/EstacaoPirata/Entities/Mobs/NPCs/changeling.yml
@@ -1,46 +1,0 @@
-- type: entity
-  name: Changeling
-  id: MobChangeling
-  parent: BaseMobOrganic
-  components:
-  - type: Hunger
-  - type: Thirst
-  - type: Perishable
-  - type: CombatMode
-  - type: InteractionPopup
-    successChance: 1
-    interactSuccessString: hugging-success-generic
-    interactSuccessSound: /Audio/Effects/thudswoosh.ogg
-    messagePerceivedByOthers: hugging-success-generic-others
-  - type: Mind
-    showExamineInfo: true
-  - type: Input
-    context: "human"
-  - type: MobMover
-  - type: InputMover
-  - type: Respirator
-    damage:
-      types:
-        Asphyxiation: 1.0
-    damageRecovery:
-      types:
-        Asphyxiation: -1.0
-  - type: Alerts
-  - type: Actions
-  - type: Eye
-  - type: CameraRecoil
-  - type: Examiner
-  - type: CanHostGuardian
-  - type: Faction
-    factions:
-    - Syndicate
-  - type: UserInterface
-    interfaces:
-    - key: enum.StoreUiKey.Key
-      type: StoreBoundUserInterface
-  - type: Store
-    categories:
-    - ChangelingAbilities
-    currencyWhitelist:
-    - Chemicals
-  - type: Changeling
