--- Resources/Prototypes/EstacaoPirata/Store/categories.yml
@@ -1,3 +1,0 @@
-- type: storeCategory
-  id: ChangelingAbilities
-  name: store-category-abilities
