--- Resources/Prototypes/EstacaoPirata/Store/currency.yml
@@ -1,9 +1,0 @@
-- type: currency
-  id: Chemicals
-  displayName: store-currency-display-chemicals
-  canWithdraw: false
-
-- type: currency
-  id: Points
-  displayName: store-currency-display-points
-  canWithdraw: false
