+++ Resources/Prototypes/Store/categories.yml
@@ -68,3 +68,7 @@
   id: RevenantAbilities
   name: store-category-abilities
 
+#changeling
+- type: storeCategory
+  id: ChangelingAbilities
+  name: store-category-abilities
