+++ Resources/Prototypes/Store/currency.yml
@@ -10,6 +10,18 @@
   displayName: store-currency-display-stolen-essence
   canWithdraw: false
 
+#changeling
+- type: currency
+  id: Chemicals
+  displayName: store-currency-display-chemicals
+  canWithdraw: false
+
+- type: currency
+  id: Points
+  displayName: store-currency-display-points
+  canWithdraw: false
+
+
 #debug
 - type: currency
   id: DebugDollar
