+++ Resources/Textures/Interface/Actions/ling_actions.rsi/arm_blade.png
