+++ Resources/Textures/Interface/Actions/ling_actions.rsi/dna_sting.png
