+++ Resources/Textures/Interface/Actions/ling_actions.rsi/fleshmend.png
