+++ Resources/Textures/Interface/Actions/ling_actions.rsi/ling_absorb.png
