+++ Resources/Textures/Interface/Actions/ling_actions.rsi/reviving_stasis.png
