+++ Resources/Textures/Interface/Actions/ling_actions.rsi/transform.png
